Restructuring and adding router
Moved home code to its own page
Getting ready for other pages using router
